Resume session at scheduled time
